Integrated password reset routes into main FastAPI application - Added /forgot-password and /reset-password endpoints to app router Promoted backend_server.py to main.py: - Renamed file and registered password reset routes - Verified /auth and root routes are live - Ready to run with: uvicorn backend.main:app --reload
